// src/app/common/model/old-data/old-order-item.interface.ts
export interface OldOrderItemInterface {
	id?: number;
    name?: string;
    category?: string;
    showall?: number;
    colors?: string;
    colorhex?: string;
    height?: number;
    width?: number;
    length?: number;
    thickness?: number;
    size?: string;
    growth?: string;
    age?: string;
    gender?: number;
    fabric?: string;
    fulfillment?: string;
    description?: string;
    price?: number;
    quantity?: number;
    availability?: number;
    availabilityinfo?: string;
    novelty?: number;
    sale?: number;
    promotion?: number;
    thumb?: string;
    images?: string;
    published?: number;
    select?: {
		size?: string;
		color?: string;
	};
    qty?: number;
}